simplify svelte example
